Game Class implementation
The Game class holds different methods to run the game loop. An instance of the game class it's created in the Main loop and it's not destroyed until user press the ESC key.

Member Functions:
void Initialize(); It initialises the game loop triggering a bool.
void Run();  It triggers the Setup, ProcessInput, Update and Output functions.
void Setup(); it will hold entities.
void ProcessInput(); it processes the Player's input.
void Update(); it will update entities.
void Output(); it will display stuff on screen.
void Stop(); it will kill the application.